+++ src/orquestra/sdk/_base/_driver/_client.py
@@ -242,7 +242,7 @@ class DriverClient:
 
     # --- queries ---
 
-    # ---- Worklow Defs ----
+    # ---- Workflow Defs ----
 
     def create_workflow_def(
         self,