+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 🔥 *Features*
 
+* New API `WorkflowRun.get_artifact()` and `WorkflowRun.get_artifact_serialized()` to get single task output
+
 🧟 *Deprecations*
 
 👩‍🔬 *Experimental*